Add Validator menu item and view with drag-and-drop control. (#1)
Add a Validator item to the top-level menu, as in the wire frame. Following a tutorial on drag-and-drop file upload in MVC, add a ValidationController, an action, and a view that presents a page with a simple drag-and-drop control.

The next step will be to invoke validation when you drag a file onto the control.

This validation page has a bug -- when I drag a .sarif file onto it, it pops the message FileNotFound. I'll address in a separate PR.