Introduce Web API ValidationController (#5)
In this change, we introduce a Web API controller to perform validation. **It does not yet actually perform the validation!** The purpose of this PR is to set up the connection between the UI and the validation logic, and to get a unit testing infrastructure in place.

We renamed the existing `ValidationController` (which displays the **TEMPORARY** validation UI) to `ValidationUiController`, and adjusted the routing accordingly.

Now when you drop a file on the filedrop area, `ValidationUiController` saves the file, then sends a request to the `ValidationController` found at `http://<root of web site>/api/Validation`. The `ValidationController` performs the validation (**not yet!**), then sends a response back to the UI controller.

The `ValidationUiController` extracts what it needs from its `Controller` base class, then delegates the real work to the `ValidationUiService`, which uses dependency injection to ensure that all the code from the service level on down is unit testable. Only the thin layer of code in `ValidationUiController` is not unit testable.

There is no code yet to display the response on the UI. I leave this to @EasyRhinoMSFT to figure out since he's our web guy.

Thanks to @EasyRhinoMSFT  for the code sample showing me how the UI controller can send a request to the web controller.