Fix #119: Online validator doesn't report all results
The problem is that the web site runs the validator with --config github.config.xml, which it gets from the SARIF SDK. But that config is intended to be used when the only thing you care about is that GitHub will accept your SARIF. As such, it disables many of the note-level rules.

The fix is to use a web-site specific configuration file that turns on the GitHub rules without turning off any other rules.